Update documentation file not to use `*` since it is causing duplicate tags.
Since `*` is used to indicate jump tags in nvim text documentation, errors are
triggered when building the tags. This update fixes that. It may not be your
preferred solution, but it is a solution. The error appears on installation of the
plugin. You can find it most easily with:

    nvim -u NONE -E -R --headless +'helptags doc' +q

There were four tag conflicts, but I went ahead and replaced all instances where I
think your intent with the stars was emphasis with backticks.